+++ waves/_tests/test_parameter_generators.py
@@ -551,30 +551,6 @@ class TestParameterGenerator:
         """
         kwargs = {"sets": sets}
         for output_file_type in _settings._allowable_output_file_types:
-            # VVV TODO: Remove dry run generator init test VVV
-            # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-            WriteParameterGenerator = DummyGenerator(
-                schema,
-                output_file_template=template,
-                output_file_type=output_file_type,
-                overwrite=overwrite,
-                dry_run=dry_run,
-                **kwargs,
-            )
-            with (
-                patch("waves.parameter_generators.ParameterGenerator._write_meta"),
-                patch("builtins.open", mock_open()) as mock_file,
-                patch("sys.stdout.write") as stdout_write,
-                patch("xarray.Dataset.to_netcdf") as xarray_to_netcdf,
-                patch("pathlib.Path.is_file", side_effect=is_file),
-                patch("pathlib.Path.mkdir"),
-            ):
-                WriteParameterGenerator.write()
-                mock_file.assert_not_called()
-                xarray_to_netcdf.assert_not_called()
-                assert stdout_write.call_count == stdout_calls
-            # ^^^ TODO: Remove dry run generator init test ^^^
-
             WriteNoDryRun = DummyGenerator(
                 schema,
                 output_file_template=template,
@@ -595,36 +571,29 @@ class TestParameterGenerator:
                 xarray_to_netcdf.assert_not_called()
                 assert stdout_write.call_count == stdout_calls
 
-    # TODO: Remove dry run from test case construction
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
     # fmt: off
-    init_write_files = {# schema, template, overwrite, dry_run,          is_file, sets, files  # noqa: E261,E721
-        "template-1":  (      {},    "out",     False,   False,          [False],    1,     1),  # noqa: E241,E201
-        "template-2":  (      {},    "out",     False,   False,   [False, False],    2,     2),  # noqa: E241,E201
-        "template-3":  (      {},    "out",     False,   False,   [ True,  True],    2,     0),  # noqa: E241,E201
-        "template-4":  (      {},    "out",     False,   False,   [ True, False],    2,     1),  # noqa: E241,E201
-        "overwrite-2": (      {},    "out",      True,   False,   [False, False],    2,     2),  # noqa: E241,E201
-        "overwrite-3": (      {},    "out",      True,   False,   [ True,  True],    2,     2),  # noqa: E241,E201
-        "overwrite-4": (      {},    "out",      True,   False,   [ True, False],    2,     2),  # noqa: E241,E201
+    init_write_files = {# schema, template, overwrite,        is_file, sets, files  # noqa: E261,E721
+        "template-1":  (      {},    "out",     False,        [False],    1,     1),  # noqa: E241,E201
+        "template-2":  (      {},    "out",     False, [False, False],    2,     2),  # noqa: E241,E201
+        "template-3":  (      {},    "out",     False, [ True,  True],    2,     0),  # noqa: E241,E201
+        "template-4":  (      {},    "out",     False, [ True, False],    2,     1),  # noqa: E241,E201
+        "overwrite-2": (      {},    "out",      True, [False, False],    2,     2),  # noqa: E241,E201
+        "overwrite-3": (      {},    "out",      True, [ True,  True],    2,     2),  # noqa: E241,E201
+        "overwrite-4": (      {},    "out",      True, [ True, False],    2,     2),  # noqa: E241,E201
     }
     # fmt: on
 
-    # TODO: Remove dry run from test case construction
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
     @pytest.mark.parametrize(
-        "schema, template, overwrite, dry_run, is_file, sets, files",
+        "schema, template, overwrite, is_file, sets, files",
         init_write_files.values(),
         ids=init_write_files.keys(),
     )
-    # TODO: Remove dry run from test case construction
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-    def test_write_yaml(self, schema, template, overwrite, dry_run, is_file, sets, files):
+    def test_write_yaml(self, schema, template, overwrite, is_file, sets, files):
         """Check for conditions that should result in calls to builtins.open
 
         :param str schema: placeholder string standing in for the schema read from an input file
         :param str template: user supplied string to be used as a template for output file names
         :param bool overwrite: overwrite existing files
-        :param bool dry_run: skip file write, but show file name and associated contents that would ahve been written
         :param list is_file: test specific argument mocks changing output for pathlib.Path().is_file() repeat calls
         :param int sets: test specific argument for the number of sets to build for the test
         :param int files: integer number of files that should be written
@@ -635,9 +604,6 @@ class TestParameterGenerator:
             output_file_template=template,
             output_file_type="yaml",
             overwrite=overwrite,
-            # TODO: Remove dry run attribute from generator init
-            # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-            dry_run=dry_run,
             **kwargs,
         )
         with (
@@ -657,9 +623,6 @@ class TestParameterGenerator:
             output_file_template=template,
             output_file_type="h5",
             overwrite=overwrite,
-            # TODO: Remove dry run attribute from generator init
-            # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-            dry_run=dry_run,
             **kwargs,
         )
         with (
@@ -674,22 +637,17 @@ class TestParameterGenerator:
             mock_write_dataset.assert_not_called()
             assert mock_write_yaml.call_count == files
 
-    # TODO: Remove dry run from test case construction
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
     @pytest.mark.parametrize(
-        "schema, template, overwrite, dry_run, is_file, sets, files",
+        "schema, template, overwrite, is_file, sets, files",
         init_write_files.values(),
         ids=init_write_files.keys(),
     )
-    # TODO: Remove dry run from test case construction
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-    def test_write_dataset(self, schema, template, overwrite, dry_run, is_file, sets, files):
+    def test_write_dataset(self, schema, template, overwrite, is_file, sets, files):
         """Check for conditions that should result in calls to ParameterGenerator._write_netcdf
 
         :param str schema: placeholder string standing in for the schema read from an input file
         :param str template: user supplied string to be used as a template for output file names
         :param bool overwrite: overwrite existing files
-        :param bool dry_run: skip file write, but show file name and associated contents that would ahve been written
         :param list is_file: test specific argument mocks changing output for pathlib.Path().is_file() repeat calls
         :param int sets: test specific argument for the number of sets to build for the test
         :param int files: integer number of files that should be written
@@ -700,9 +658,6 @@ class TestParameterGenerator:
             output_file_template=template,
             output_file_type="h5",
             overwrite=overwrite,
-            # TODO: Remove dry run attribute from generator init
-            # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-            dry_run=dry_run,
             **kwargs,
         )
         with (
@@ -724,9 +679,6 @@ class TestParameterGenerator:
             output_file_template=template,
             output_file_type="yaml",
             overwrite=overwrite,
-            # TODO: Remove dry run attribute from generator init
-            # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-            dry_run=dry_run,
             **kwargs,
         )
         with (