+++ waves/scons_extensions.py
@@ -3997,11 +3997,6 @@ def parameter_study_write(
     """
     import yaml
 
-    # TODO: move the dry run option out of the parameter generator and into the write API
-    # Check the kwargs for the dry run option instead of the parameter generator attributes
-    # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/859
-    if parameter_generator.dry_run:
-        raise RuntimeError("Parameter generator's dry run attribute is set to True")
     # TODO: Refactor write/output file logic to avoid duplication here
     if parameter_generator.output_file is not None:
         output_files = [parameter_generator.output_file]